implemented RecipeListAdapter.kt and RecipesFragment.kt

+++ app/src/main/java/com/sanmidev/mybakingapp/data/remote/model/IngredientModel.kt
@@ -11,5 +11,5 @@ data class IngredientModel(
     @Json(name = "measure")
     val measure: String, // CUP
     @Json(name = "quantity")
-    val quantity: Int // 2
+    val quantity: Double // 2
 )
